ticket-booking-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
redis==6.4.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
email-validator==2.3.0
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1